REF: organizer code to be reusable for other tools
